+++ src/apps/settings/modules/general/main/infra/index.tsx
@@ -1,5 +1,5 @@
-import { Select, Switch } from 'antd';
-import { useState } from 'react';
+import { Input, Select, Switch, Tooltip } from 'antd';
+import { ChangeEvent, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 
@@ -8,6 +8,7 @@ import { useAppDispatch } from '../../../shared/utils/infra';
 
 import { RootActions } from '../../../shared/store/app/reducer';
 import { RootSelectors } from '../../../shared/store/app/selectors';
+import { Icon } from 'src/apps/shared/components/Icon';
 
 import { LanguageList } from '../../../../../shared/lang';
 import { SettingsGroup, SettingsOption } from '../../../../components/SettingsBox';
@@ -20,6 +21,7 @@ export function General() {
 
   const autostartStatus = useSelector(RootSelectors.autostart);
   const language = useSelector(RootSelectors.language);
+  const dateFormat = useSelector(RootSelectors.dateFormat);
 
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
@@ -35,19 +37,24 @@ export function General() {
     dispatch(RootActions.setAutostart(value));
   };
 
+  const onDateFormatChange = (e: ChangeEvent<HTMLInputElement>) =>
+    dispatch(RootActions.setDateFormat(e.target.value));
+
   return (
     <>
       <SettingsGroup>
         <SettingsOption>
-          <span style={{ fontWeight: 600 }}>{t('general.startup')}</span>
+          <b>{t('general.startup')}</b>
           <Switch
             onChange={onAutoStart}
             value={!!autostartStatus}
             loading={changingAutostart || autostartStatus === null}
           />
         </SettingsOption>
+      </SettingsGroup>
+      <SettingsGroup>
         <SettingsOption>
-          <b>{t('general.language')}:</b>
+          <b>{t('general.language')}</b>
           <Select
             showSearch
             optionFilterProp="label"
@@ -55,6 +62,26 @@ export function General() {
             value={language}
             options={[...LanguageList]}
             onSelect={(value) => dispatch(RootActions.setLanguage(value))}
+          />
+        </SettingsOption>
+        <SettingsOption>
+          <div style={{ display: 'flex', alignItems: 'center', gap: '6px' }}>
+            <b>{t('general.date_format')}</b>
+            <Tooltip
+              title={
+                <a href="https://momentjs.com/docs/#/displaying/format/" target="_blank">
+                  https://momentjs.com/docs/#/displaying/format/
+                </a>
+              }
+            >
+              <Icon iconName="LuHelpCircle" />
+            </Tooltip>
+          </div>
+          <Input
+            style={{ width: '200px', maxWidth: '200px' }}
+            placeholder="YYYY-MM-DD"
+            value={dateFormat}
+            onChange={onDateFormatChange}
           />
         </SettingsOption>
       </SettingsGroup>