+++ src/apps/settings/modules/shared/store/app/selectors.ts
@@ -13,6 +13,3 @@ export const SeelenWegSelectors = SeelenWegSlice.getSelectors(RootSelectors.seel
 export const SeelenWmSelectors = SeelenManagerSlice.getSelectors(RootSelectors.windowManager);
 
 export const BorderSelectors = BorderSlice.getSelectors(SeelenWmSelectors.border);
-
-export const getMonitorSelector = (idx: number) => (state: RootState) => RootSelectors.monitors(state)[idx];
-export const getWorkspaceSelector = (idx: number, monitorIdx: number) => (state: RootState) => getMonitorSelector(monitorIdx)(state)?.workspaces[idx];